% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{augment_Bvec}
\alias{augment_Bvec}
\title{augment_Bvec
Function for augmentig the breakpoint vector Bvec as result from accepting a forward step}
\usage{
augment_Bvec(Bvec, T_p, IN, S)
}
\arguments{
\item{T_p}{(const double) vector to be inserted}

\item{IN}{(const int) index of the segment splitted}

\item{S}{(const int) number of current segments}

\item{vec}{(NumericVector&) vector with breakpoint}
}
\value{
void no output is returned because the input (vec) is passed by reference and hence modified
}
\description{
augment_Bvec
Function for augmentig the breakpoint vector Bvec as result from accepting a forward step
}
