% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{augment_vecM}
\alias{augment_vecM}
\title{augment_vecM
Function for augment/updating the state vector Mvec as resulting from accepting a forward step}
\usage{
augment_vecM(vec, left_value, right_value, IN, S)
}
\arguments{
\item{vec}{(NumericVector&) vector to be augmented}

\item{left_value}{(const double) state value of the new left segment}

\item{right_value}{(const double) state value of the new right segment}

\item{IN}{(const int) index of the  segment splitted (range is (1,S) )}

\item{S}{(const int) number of current segments}
}
\value{
void no output is returned because the input (vec) is passed by reference and hence modified
}
\description{
augment_vecM
Function for augment/updating the state vector Mvec as resulting from accepting a forward step
}
