% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{augment_vecZQFV}
\alias{augment_vecZQFV}
\title{augment_vecZQFV
Function for augment/updating the state vectors (Zvec, Vvec, Fvec) as resulting from accepting a forward step}
\usage{
augment_vecZQFV(vec, left_value, right_value, IN, S)
}
\arguments{
\item{vec}{(IntegerVector&) vector to be augmented}

\item{left_value}{(const int) state value of the new left segment}

\item{right_value}{(const int) state value of the new right segment}

\item{IN}{(const int) index of the the  segment splitted(range is (1,S) )}

\item{S}{(const int) number of current segments}
}
\value{
void no output is returned because the input (vec) is passed by reference and hence modified
}
\description{
augment_vecZQFV
Function for augment/updating the state vectors (Zvec, Vvec, Fvec) as resulting from accepting a forward step
}
