% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints_sampling.R
\name{breakpoints_sampling}
\alias{breakpoints_sampling}
\title{Function for generating a sample of Breakpoints inside an observational interval}
\usage{
breakpoints_sampling(start_point, end_point, breakpoint_list, sample_size)
}
\arguments{
\item{start_point}{start point for the segment (t_star)}

\item{end_point}{end point of the observational interval}

\item{breakpoint_list}{list with breakpoints vectors sampled during the RJSMCMC}

\item{sample_size}{number of samples to generate}
}
\value{
list called output with following objects
\describe{
\item{logdens}{vector with log density value for  the generate breakpoint vector, for each sample }
\item{ S.samp}{list of vectors with the generated breakpoints in each sample }
}
}
\description{
Function for generating breakpoints from a kernel density inside an interval
}
