% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{collapsed_full_conditional_Q}
\alias{collapsed_full_conditional_Q}
\title{collapsed_full_conditional_Q
Function for sampling the Q  state for a  new non-empty segment from its  collpsed full conditional and/or evaluate it at the sampled point
It is "collapsed" because the likelihood part is not Gamma(M,alpha_q,alpha_q/mu_q) but
shifted Neg-Bin (N;alpha_q,L/(alpha_q/mu_q + L)). M is integrated out because not needed}
\usage{
collapsed_full_conditional_Q(
  Q,
  W,
  count,
  LB,
  UB,
  alphavec,
  muvec,
  probvec_Q,
  minimum_n,
  open_segment = FALSE,
  end_point = 0,
  sample_Q = FALSE
)
}
\arguments{
\item{Q}{(const int&) Value of the current Q if present; if not set it 0.}

\item{W}{(const int&) Number of levels for the Q state}

\item{count}{(const double&) number of observations within the selected segment}

\item{LB}{(const double&) lower bound of the selected segment}

\item{UB}{(const double&) upper bound of the selected segment}

\item{alphavec}{(const NumericVector&) Shape parameters ruling the rate of occurrence of segment for the different levels of the Q state}

\item{muvec}{(const NumericVector&) Mean parameters ruling the rate of occurrence of segment for the different levels of the Q state}

\item{probvec_Q}{(const NumericVector&) Probability mass for the Q state distribution}

\item{minimum_n}{minimum number of observations that must be observed in a non-empty segment}

\item{open_segment}{(const bool) if false (by default) the segment is considered totally observed. If true the semgent is considered partially observed; this affects the likelihood computation}

\item{end_point}{(double, default is 0.0) to be used for partially observed segments; declares the point where the stops to be observed}

\item{sample_Q}{( const bool=false) if true a new Q is sampled and the full conditional is evaluated at the new value. If false (by default) the full condititional is evaluated at the input Q}
}
\value{
A list containting 2 elements:
\describe{
\item{Q_new}{int, the sample Q state for the selected segment}
\item{eval_densQ}{double, value of the log full conditional for either the new sampled Q or the current Q inputed ( depending on sample_Q)}
}
}
\description{
collapsed_full_conditional_Q
Function for sampling the Q  state for a  new non-empty segment from its  collpsed full conditional and/or evaluate it at the sampled point
It is "collapsed" because the likelihood part is not Gamma(M,alpha_q,alpha_q/mu_q) but
shifted Neg-Bin (N;alpha_q,L/(alpha_q/mu_q + L)). M is integrated out because not needed
}
