% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_lambdavec.R
\name{compute_lambdavec}
\alias{compute_lambdavec}
\title{compute_lambdavec
function for generating the lambda parameters, a matrix where each row is the probability distribution
for the log messages (here we use LETTERS) for each V state assumed}
\usage{
compute_lambdavec(name, prob_name)
}
\arguments{
\item{name}{vector containing the LETTTERS we want to assign a specific probability}

\item{prob_name}{vector with assigned probabilities for the LETTERS specified in \code{name}}
}
\value{
matrix with probability vector for each LETTER, in each V state
}
\description{
compute_lambdavec
function for generating the lambda parameters, a matrix where each row is the probability distribution
for the log messages (here we use LETTERS) for each V state assumed
}
