% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_posterior}
\alias{compute_posterior}
\title{Function for updating the posterior distribution of state variables}
\usage{
compute_posterior(
  num_discr_intervals,
  num_particles,
  state_container,
  num_states,
  weight_vec
)
}
\arguments{
\item{num_discr_intervals}{(int) number of discretization intervals}

\item{num_particles}{(int) number of particles}

\item{state_container}{(IntegerMatrix) with inferred state values for each discretization interval, for each particle}

\item{num_states}{(int) number of states}

\item{weight_vec}{(NumericVector) weights assigned to each particle}
}
\description{
Function for updating the posterior distribution of state variables
}
