% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_posterior_breakpoint}
\alias{compute_posterior_breakpoint}
\title{Function for computing the posterior distribution of breakpoint locations}
\usage{
compute_posterior_breakpoint(
  num_discr_intervals,
  num_particles,
  state_container_B,
  weight_vec
)
}
\arguments{
\item{num_discr_intervals}{(int) number of discretization intervals}

\item{num_particles}{(int) number of particles}

\item{state_container_B}{(NumericMatrix) matrix indicating breakpoint presence: 1 if particle j has a breakpoint in interval i, 0 otherwise}

\item{weight_vec}{(NumericVector) weights assigned to each particle (should sum to 1)}
}
\value{
NumericVector posterior probability that a breakpoint occurs in each discretization interval
}
\description{
Function for computing the posterior distribution of breakpoint locations
}
