% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_findInterval}
\alias{cpp_findInterval}
\title{Function to assign each element of a vector "x" to one of N  unique intervals defined through the
vector of breakpoint B; B\link{j}<=x<\link{j+1} j=1...N+1. The out is "j"}
\usage{
cpp_findInterval(x, breaks)
}
\arguments{
\item{x}{NumericVector. Vector  whose element need to ble allocated}

\item{breaks}{NumericVector. Vector containing the N+1 breakpoints}
}
\value{
IntegerVector having the same size as x. Each element contains the index referring to the interval x\link{i} belongs to
}
\description{
Function to assign each element of a vector "x" to one of N  unique intervals defined through the
vector of breakpoint B; B\link{j}<=x<\link{j+1} j=1...N+1. The out is "j"
}
