% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data_function.R
\name{data_simulation}
\alias{data_simulation}
\title{function for generating simulation data
This function, given a set of hyper-parameters, generates a complete dataset. The generating
model is the model of paper 1 but assuming 3 states (Z,Q,V) for length, rate and composition
of the segments, respectively.}
\usage{
data_simulation(
  probvec_Z,
  probvec_V,
  probvec_Q,
  probvec_F,
  P0,
  alphavec,
  muvec,
  keyvec,
  etavec,
  key0vec,
  eta0vec,
  lambdamat,
  K,
  U,
  W,
  Bmax,
  seed,
  min_obs = 3
)
}
\arguments{
\item{probvec_Z}{Probability vector for state Z}

\item{probvec_V}{Probability vector for state V}

\item{probvec_Q}{Probability vector for state Q}

\item{probvec_F}{Probability vector for state F}

\item{P0}{Probability to observe an empty segment after a non-empty one}

\item{alphavec}{Shape parameters for the rate of the non-empty segments}

\item{muvec}{Mean parameters for the rate of the non-empty segments}

\item{keyvec}{Shape parameters for the length of the non-empty segments}

\item{etavec}{Mean parameters for the length of the non-empty segments}

\item{key0vec}{Shape parameters for the length of the empty segments}

\item{eta0vec}{Mean parameters for the length of the empty segments}

\item{lambdamat}{Probability vectors for the message composition of the V states}

\item{K}{Number of levels for state Z}

\item{U}{Number of levels for state V}

\item{W}{Number of levels for state Q}

\item{Bmax}{Final breakpoint}

\item{seed}{Seed for reproducibility}

\item{min_obs}{Minimum number of observations allowed in a segment, default is 3}
}
\value{
list with following objects
\describe{
\item{Bvec}{ Vector with generated changepoints}
\item{Vvec}{Vector with generated V's - ruling the message distribution in the non-empty segments}
\item{Zvec}{Vector with generated Z's - ruling the length of  non-empty segments}
\item{Qvec}{Vector with generated Q's - ruling the rate of messages in non-empty segments}
\item{Fvec}{Vector with generated F's - ruling the length of the empty segments}
\item{Mvec}{Vector with generated M's - rates of messages in non-empty segments}
\item{Nvec}{Vector with generated N's - counts of observations in non-empty segments}
\item{Tvec}{Vector with genetated time stamps}
\item{Yvec}{Vector with genetated messages}
\item{probvec_Z}{vector with probabilities for the levels of state Z}
\item{probvec_V}{vector with probabilities for the levels of state V}
\item{probvec_F}{vector with probabilities for the levels of state F}
\item{probvec_Q}{vector with probabilities for the levels of state Q}
\item{P0}{Probability to observe an empty segment after a non-empty one}
\item{alphavec}{Shape parameters for the rate of the non-empty segments}
\item{muvec}{Mean parameters for the rate of the non-empty segments}
\item{keyvec}{Shape parameters for the length of the non-empty segments}
\item{key0vec}{Shape parameters for the lenght of the empty segments }
\item{etavec}{Mean parameters for the length of the non-empty segments}
\item{eta0vec}{ Mean parameters for the lenght of the empty segments}
\item{lambdamat}{Probability vectors for the message composition of the V states}
\item{K}{Number of levels for state Z}
\item{U}{Number of levels for state V}
\item{W}{Number of levels for state Q}
\item{num_logs}{number of different possible messages}
\item{minimum_n}{minimum number of messages allowed in a segment}
}
}
\description{
function for generating simulation data
This function, given a set of hyper-parameters, generates a complete dataset. The generating
model is the model of paper 1 but assuming 3 states (Z,Q,V) for length, rate and composition
of the segments, respectively.
}
