% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{english_words_low_rate_08}
\alias{english_words_low_rate_08}
\title{English words low-rate 08 dataset}
\format{
Same as \code{english_words}; a list with \code{Tvec}, \code{Yvec}, \code{Bvec}, \code{Vvec}, and all parameters.
}
\usage{
english_words_low_rate_08
}
\description{
Same structure as \code{english_words} but with lower message rates 0.8 probability for the messages
}
\keyword{datasets}
