% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_index}
\alias{extract_index}
\title{extract_index
Function for extracting the first and last index elements of an ordered vector
sastisfying a constraint specified by a lower and upper bound}
\usage{
extract_index(x, lower, upper)
}
\arguments{
\item{x}{(const NumericVector&) the original vector}

\item{lower}{(const double), the lower bound (included in the subsetting)}

\item{upper}{(const double), the lower bound}
}
\value{
3-element vector with:
\describe{
\item{element \link{0}}{length of the subsetted vector}
\item{element \link{1}}{index of the first element of the subsetted vector }
\item{element \link{2}}{index of the last element of the subsetted vector. If the last index is "-1" then no elements sastisfy the constraint}
}
}
\description{
extract_index
Function for extracting the first and last index elements of an ordered vector
sastisfying a constraint specified by a lower and upper bound
}
