% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{full_conditional_F}
\alias{full_conditional_F}
\title{full_conditional_F
Function for sampling the F state for a  new empty segment from its full conditional and/or evaluate it at the sampled point}
\usage{
full_conditional_F(F, L, key0vec, eta0vec, probvec_F, sample_F = FALSE)
}
\arguments{
\item{F}{(const int&) Value of the current F if present; if not set it 0.}

\item{L}{(const double&) Length of the selected segment}

\item{key0vec}{(const NumericVector&) Shape parameters ruling the length of empty segment for the different levels of the F state}

\item{eta0vec}{(const NumericVector&) Mean parameters ruling the length of empty segment for the different levels of the F state}

\item{probvec_F}{(const NumericVector&) Probability mass for the F state distribution}

\item{sample_F}{( const bool=false) if true a new F is sampled and the full conditional is evaluated at the new value. If false (by default) the full condititional is evaluated at the input F}
}
\value{
A list containting 2 elements:
\describe{
\item{F_new}{int, the sample F state for the selected segment}
\item{eval_densF}{double, value of the  log full conditional for either the new F sampled of the current F inputed ( depending on sample_F)}
}
}
\description{
full_conditional_F
Function for sampling the F state for a  new empty segment from its full conditional and/or evaluate it at the sampled point
}
