% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{full_conditional_V}
\alias{full_conditional_V}
\title{full_conditional_V
Function for sampling the V  state for a  new non-empty segment from its full conditional and/or evaluate it at the sampled point}
\usage{
full_conditional_V(
  V,
  U,
  num_logs,
  T_seg,
  Y_seg,
  LB,
  UB,
  lambdamat,
  probvec_V,
  V_left,
  P0,
  end_point = 0,
  open_segment = FALSE,
  sample_V = FALSE
)
}
\arguments{
\item{V}{(const int&) Value of the current V if present; if not set it 0.}

\item{U}{(const int&) Number of levels for the V state}

\item{num_logs}{(const int&) Total number of unique messages that can be observed}

\item{T_seg}{(const NumericVector&) Vector with time stamps within the update interval}

\item{Y_seg}{(const IntegerVector&) Vector with the messages within the update interval}

\item{LB}{(const double&) lower bound of the selected segment}

\item{UB}{(const double&) upper bound of the selected segment}

\item{lambdamat}{(const NumericMatrix&) probability weights ruling the message composition of the segments for the different levels of the V state}

\item{probvec_V}{(const NumericVector&) Probability mass for the V state distribution}

\item{V_left}{V value for the previous segment}

\item{P0}{(const double&) Probability to observe an empty segment after a non-empty one}

\item{end_point}{(double, default is 0.0) to be used for partially observed segments; declares the point where the stops to be observed}

\item{open_segment}{(const bool, default is false) if false  the segment is considered totally observed. If true the semgent is considered partially observed; this affects the likelihood computation}

\item{sample_V}{( const bool=false) if true a new V is sampled and the full conditional is evaluated at the new value. If false (by default) the full condititional is evaluated at the input V}
}
\value{
A list containting 2 elements:
\describe{
\item{V_new}{int, the sample Vstate for the selected segment or the one inputed}
\item{eval_densV}{double, value of the log full conditional for either the new sampled V or the current Q inputed ( depending on sample_V)}
}
}
\description{
full_conditional_V
Function for sampling the V  state for a  new non-empty segment from its full conditional and/or evaluate it at the sampled point
}
