% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{full_conditional_Z}
\alias{full_conditional_Z}
\title{full_conditional_Z
Function for sampling the Z state for a  new non-empty segment from its full conditional and/or evaluate it at the sampled point}
\usage{
full_conditional_Z(Z, K, L, keyvec, etavec, probvec_Z, sample_Z = FALSE)
}
\arguments{
\item{Z}{(const int&) Value of the current Z if present; if not set it 0.}

\item{K}{(const int&) Number of levels for the Z state}

\item{L}{(const double&) Length of the selected segment}

\item{keyvec}{(const NumericVector&) Shape parameters ruling the length of segment for the different levels of the Z state}

\item{etavec}{(const NumericVector&) Mean parameters ruling the length of segment for the different levels of the Z state}

\item{probvec_Z}{(const NumericVector&) Probability mass for the Z state distribution}

\item{sample_Z}{( const bool=false) if true a new Z is sampled and the full conditional is evaluated at the new value. If false (by default) the full condititional is evaluated at the input Z}
}
\value{
A list containting 2 elements:
\describe{
\item{Z_new}{int, the sample Z state for the selected segment}
\item{eval_densZ}{double, value of the  log full conditional for either the new Z sampled of the current Z inputed ( depending on sample_Z)}
}
}
\description{
full_conditional_Z
Function for sampling the Z state for a  new non-empty segment from its full conditional and/or evaluate it at the sampled point
}
