% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\title{Function for extrapolating the results
In this function the approximated posterior distributions are extracted.
The observational time interval is discretized by regularly-spaced time points
defined through the parameter \code{interval_length}. For each time point
the posterior distribution of the dynamic states V, Z, Q and F is computed.}
\usage{
get_results(
  start_point,
  end_point,
  num_particles,
  interval_length,
  breakpoints_list,
  V_state_list,
  Z_state_list,
  Q_state_list,
  F_state_list,
  num_states_V,
  num_states_Z,
  num_states_Q,
  num_states_F
)
}
\arguments{
\item{start_point}{starting point of the update interval}

\item{end_point}{ending point of the update interval}

\item{num_particles}{the number of particles generated in SMC}

\item{interval_length}{the chosen length for spacing each discretization time point}

\item{breakpoints_list}{list with inferred breakpoints inside the update interval, for all particles}

\item{V_state_list}{list with inferred V state at each breakpoint, for all particles}

\item{Z_state_list}{list with inferred Z state at each breakpoint, for all particles}

\item{Q_state_list}{list with inferred Q state at each breakpoint, for all particles}

\item{F_state_list}{list with inferred F state at each breakpoint, for all particles}

\item{num_states_V}{number of states for V}

\item{num_states_Z}{number of states for Z}

\item{num_states_Q}{number of states for Q}

\item{num_states_F}{number of states for F}
}
\description{
Function for extrapolating the results
In this function the approximated posterior distributions are extracted.
The observational time interval is discretized by regularly-spaced time points
defined through the parameter \code{interval_length}. For each time point
the posterior distribution of the dynamic states V, Z, Q and F is computed.
}
