% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{incremental_weight}
\alias{incremental_weight}
\title{Incremental_weight
Function for computing the log incremental weight for a particle}
\usage{
incremental_weight(
  T_seg,
  Y_seg,
  V,
  Z,
  Q,
  F,
  U,
  K,
  W,
  B_last,
  B1,
  num_logs,
  lambdamat,
  keyvec,
  etavec,
  key0vec,
  eta0vec,
  alphavec,
  muvec,
  probvec_V,
  probvec_Z,
  probvec_Q,
  probvec_F,
  V_last_complete,
  P0,
  minimum_n,
  t_star,
  end_point = 0,
  open_segment = FALSE
)
}
\arguments{
\item{T_seg}{Vector with time stamps within the update interval}

\item{Y_seg}{Vector with the messages within the update interval}

\item{V}{(const int&) Value of the current V if present; if not set it 0.}

\item{Z}{(const int&) Value of the current Z if present; if not set it 0.}

\item{Q}{(const int&) Value of the current Q if present; if not set it 0.}

\item{F}{(const int&) Value of the current F if present; if not set it 0.}

\item{U}{(const int&) Number of levels for the V state}

\item{K}{(const int&) Number of levels of the Z state}

\item{W}{(const int&) Number of levels for the Q state}

\item{B_last}{Last simulated changepoint WITHIN the Update Interval, at the previous SMC iteration}

\item{B1}{(const double&) first generate changepoint, at the current SMC iteration}

\item{num_logs}{(const int&) Total number of unique messages that can be observed}

\item{lambdamat}{(const NumericMatrix&) probability weights ruling the message composition of the segments for the different levels of the V state}

\item{keyvec}{(const NumericVector&) Shape parameters ruling the length of non-empty segments in different levels of Z state}

\item{etavec}{(const NumericVector&) Mean parameters ruling the length of non-empty segments in different levels of Z state}

\item{key0vec}{(const NumericVector&) Shape parameters ruling the length of empty segments in different levels of F state}

\item{eta0vec}{(const NumericVector&) Mean parameters ruling the length of empty segments in different levels of F state}

\item{alphavec}{(const NumericVector&) Shape parameters ruling the rate of occurrence of segment for the different levels of the Q state}

\item{muvec}{(const NumericVector&) Mean parameters ruling the rate of occurrence of segment for the different levels of the Q state}

\item{probvec_V}{(const NumericVector&) Probability mass for the V state distribution}

\item{probvec_Z}{(const NumericVector&) probability mass for the Z state}

\item{probvec_Q}{(const NumericVector&) Probability mass for the Q state distribution}

\item{probvec_F}{(const NumericVector&) probability mass for the F state}

\item{V_last_complete}{(const int&) V value for the last "closed" segment generated during the previous SMC iteration}

\item{P0}{(const double&) Probability to observe an empty segment after a non-empty one}

\item{minimum_n}{(const int&) minimum number of observations that must be observed in a non-empty segment}

\item{t_star, }{(double) estimate of the last simulated breakpoint}

\item{end_point}{(double, default is 0.0) to be used for partially observed segments; declares the point where the segment stops to be observed}

\item{open_segment}{(const bool, default is false) if false  the segment is considered totally observed. If true the segment is considered partially observed; this affects the likelihood computation}
}
\description{
Incremental_weight
Function for computing the log incremental weight for a particle
}
