% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initial_breakpoints}
\alias{initial_breakpoints}
\title{initial_breakpoints
Function to initialize the changepoint vector during the first iteration of the RJMCMC}
\usage{
initial_breakpoints(
  T_seg_orig,
  t_star,
  start_point,
  end_point,
  minimum_n,
  max_range,
  min_seg_length
)
}
\arguments{
\item{T_seg_orig}{Vector with time stamps within the update interval}

\item{t_star}{Estimate of the latest changepoint simulated when performing the RJMCMC during the previous iteration of the SMC (precedent update interval)}

\item{start_point}{Left bound of the current update interval}

\item{end_point}{Right bound of the current update interval}

\item{minimum_n}{minimum number of observation allowed in a non-empty segment}

\item{max_range}{(const double&) time interval in which the last changepoint is allowed to be place after the end point (end_point;end_point+1)}

\item{min_seg_length}{(const double) minimum length allowed for a segment to be valid}
}
\value{
NumericVector with the generated changepoints. Starting with t_star, ending with B_k > UB
}
\description{
initial_breakpoints
Function to initialize the changepoint vector during the first iteration of the RJMCMC
}
