% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initial_state}
\alias{initial_state}
\title{initial_state
Function for generating the initial state values in each segment}
\usage{
initial_state(
  T_seg,
  Y_seg,
  S,
  Bvec,
  end_point,
  probvec_V,
  probvec_Q,
  probvec_Z,
  probvec_F,
  lambdamat,
  U,
  W,
  K,
  alphavec,
  muvec,
  keyvec,
  etavec,
  key0vec,
  eta0vec,
  P0,
  num_logs,
  empty_mix,
  minimum_n
)
}
\arguments{
\item{T_seg}{Vector with time stamps within the update interval}

\item{Y_seg}{Vector with the messages within the update interval}

\item{S}{Number of segments within the update interval}

\item{Bvec}{Position of the Changepoints simulated within the update interval}

\item{end_point}{(const double&) end point of the update interval}

\item{probvec_V}{probability mass for the V state}

\item{probvec_Q}{probability mass for the Q state}

\item{probvec_Z}{probability mass for the Z state}

\item{probvec_F}{probability mass for the F state}

\item{lambdamat}{matrix U*num_logs with probability mass for the messages in each V state}

\item{U}{Number of levels of the V state}

\item{W}{Number of levels of the Q state}

\item{K}{Number of levels of the Z state}

\item{alphavec}{Shape parameters ruling the rate of occurrence in different levels of Z state (non-empty segments)}

\item{muvec}{Mean parameters ruling the rate of occurrence in different levels of Z state}

\item{keyvec}{Shape parameters ruling the length of non-empty segments in different levels of Z state}

\item{etavec}{Mean parameters ruling the length of non-empty segments in different levels of Z state}

\item{key0vec}{Shape parameters ruling the length of empty segments in different levels of F state}

\item{eta0vec}{Mean parameters ruling the length of empty segments in different levels of F state}

\item{P0}{(const double&) Probability to observe an empty segment after a non-empty one}

\item{num_logs}{Total number of different messages that can be observed (dictionary)}

\item{empty_mix}{Logical. Equal "true" if the empty segments are a mixture of the states}

\item{minimum_n}{minimum number of observations that must be observed in a non-empty segment}
}
\value{
A list containting 6 elements:
\describe{
\item{Vvec}{IntegerVector,  initial V state for each segment }
\item{Zvec}{IntegerVector,  initial Z state for each segment }
\item{Qvec}{IntegerVector,  initial Q state for each segment }
\item{Fvec}{IntegerVector,  initial F state for each segment }
\item{Mvec}{NumericVector,  initial M (rate) for each segment }
\item{Nvec}{NumericVector, count of observations in each segment }
}
}
\description{
initial_state
Function for generating the initial state values in each segment
}
