% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iteration0_RJMCMC}
\alias{iteration0_RJMCMC}
\title{Function to initialize the MCMC in RJMCMC_SMC
Function generating the first iteration of the MCMC}
\usage{
iteration0_RJMCMC(
  T_seg,
  Y_seg,
  minimum_n,
  start_point,
  end_point,
  t_star,
  K,
  W,
  U,
  empty_mix,
  probvec_V,
  probvec_Z,
  probvec_Q,
  probvec_F,
  alphavec,
  muvec,
  keyvec,
  etavec,
  key0vec,
  eta0vec,
  lambdamat,
  P0,
  num_logs,
  max_range,
  Smax
)
}
\arguments{
\item{T_seg}{Vector with time stamps within the update interval}

\item{Y_seg}{Vector with the messages within the update interval}

\item{minimum_n}{(int) minimum allowed number of observations in a active segments}

\item{start_point, }{(double)   the start point of the update interval}

\item{end_point, }{(double)   the end point of the update interval}

\item{t_star, }{(double) estimate of the last simulated breakpoint}

\item{K}{(int) number of Z states}

\item{W}{(int) Number of levels for the Q state}

\item{U}{(int) Number of V states}

\item{empty_mix}{(bool) If True (default) 2 kind of empty segments are assumed}

\item{probvec_V}{probability mass for the V state}

\item{probvec_Z}{probability mass for the Z state}

\item{probvec_Q}{probability mass for the Q state}

\item{probvec_F}{probability mass for the F state}

\item{alphavec}{Shape parameters ruling the rate of occurrence in different levels of Z state (non-empty segments)}

\item{muvec}{Mean parameters ruling the rate of occurrence in different levels of Z state}

\item{keyvec}{Shape parameters ruling the length of non-empty segments in different levels of Z state}

\item{etavec}{Mean parameters ruling the length of non-empty segments in different levels of Z state}

\item{key0vec}{Shape parameters ruling the length of empty segments in different levels of F state}

\item{eta0vec}{Mean parameters ruling the length of empty segments in different levels of F state}

\item{lambdamat}{(const NumericMatrix&) each row v is the full prob vector for the messages of state V=v}

\item{P0}{(const double&) Probability to observe an empty segment after a non-empty one}

\item{num_logs}{Total number of different messages that can be observed (dictionary)}

\item{max_range}{(const double&) time interval in which the last changepoint is allowed to be place after the end point (end_point;end_point+1)}

\item{Smax}{maximum number of segments allowed inside the update interval}
}
\value{
A list containting 8 elements:
\describe{
\item{S}{int, number of segments generarted}
\item{Bvec}{NumericVector, hosts the changepoints generated (inlcuding t_star, tau_k (last breakpoint after end_point)) }
\item{Vvec}{IntegerVector,  initial V state for each segment }
\item{Zvec}{IntegerVector,  initial Z state for each segment }
\item{Qvec}{IntegerVector,  initial Q state for each segment }
\item{Fvec}{IntegerVector,  initial F state for each segment }
\item{Mvec}{NumericVector,  initial M (rate) for each segment }
\item{Nvec}{NumericVector, count of observations in each segment }
}
}
\description{
Function to initialize the MCMC in RJMCMC_SMC
Function generating the first iteration of the MCMC
}
