% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_table}
\alias{make_table}
\title{Function for creating ax frequency table.}
\usage{
make_table(X, K, inc_zero = FALSE)
}
\arguments{
\item{X}{NumericVector. Vector whose element are used to create the frequency table}

\item{K}{Integer. The number of classes -max(X) standard choice-}

\item{inc_zero}{Boolean. inc_zero==F (default) if the 0's must not be taken into account in the counting.}
}
\value{
The frequency vector (NumericVector)
}
\description{
Function for creating ax frequency table.
}
