% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RJSMCdata.R
\name{plot.RJSMCdata}
\alias{plot.RJSMCdata}
\title{function for plotting RJSMC data}
\usage{
\method{plot}{RJSMCdata}(d)
}
\arguments{
\item{d}{list containing two components,
\describe{
\item{time}  Vector of time-points of class 'Date'
\item{labels} Vector of labels, categorical (or integer)
}}
}
\value{
list with following objects
\describe{
\item{minimum_n}{minimum number of messages allowed in a segment}
}
}
\description{
function for plotting RJSMC data
}
