% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prob_approx}
\alias{prob_approx}
\title{Function for approximating the log probability of observing "n" obs in semi-segment [tau0,split_U) given
the total length [tau0,tau1)}
\usage{
prob_approx(count, split_U, alpha, mu, minimum_n, tau0, tau1, error)
}
\arguments{
\item{count}{(double& int) number of observations within the selected segment}

\item{split_U}{(const& double) point where the update interval ends}

\item{alpha}{(const& double) shape parameter for the rate}

\item{mu}{(const& double) mean parameter for the rate}

\item{minimum_n}{(const& int) minimum number of observations allowed in a non-empty segment}

\item{tau0}{(const& double) left bound of the segment}

\item{tau1}{(const& double) right bound of the segment}

\item{error}{(const& double) approximation error}
}
\value{
the value of the log density
}
\description{
Function for approximating the log probability of observing "n" obs in semi-segment [tau0,split_U) given
the total length [tau0,tau1)
}
