% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{real_data}
\alias{real_data}
\title{Real data dataset}
\format{
a list containing parameters and variables:
\describe{
\item{Tvec}{Vector containing the time stamps}
\item{Yvec}{Vector containing the messages}
\item{minimum_n}{minimum number of observations allowed in each non-empty segment}
\item{K}{number levels of the variable Z}
\item{U}{number levels of the variable V}
\item{W}{number levels of the variable Q}
\item{probvec_Z}{vector with probability distribution for variable Z}
\item{probvec_V}{vector with probability distribution for variable V}
\item{probvec_Q}{vector with probability distribution for variable Q}
\item{probvec_F}{vector with probability distribution for variable F}
\item{P0}{probability to observe an empty segment after a non-empty one}
\item{lambdamat}{matrix U*num_logs, each row contains the probability distribution for the message for each level of variable V}
\item{alphavec}{Shape parameters ruling the rate of occurrence for the non-empty segments, for each level of variable Q}
\item{muvec}{shape parameters ruling the rate of occurrence for the non-empty segments, for each level of variable Q}
\item{keyvec}{shape parameters ruling the rate of occurrence for the non-empty segments, for each level of variable Z}
\item{etavec}{mean parameters ruling the rate of occurrence for the non-empty segments, for each level of variable Z}
\item{key0vec}{shape parameters ruling the rate of occurrence for the non-empty segments, for each level of variable F}
\item{eta0vec}{mean parameters ruling the rate of occurrence for the non-empty segments, for each level of variable F}
\item{num_logs}{total number of observable messages}
\item{Bvec}{NULL (not available for real data)}
\item{Vvec}{NULL (not available for real data)}
\item{Zvec}{NULL (not available for real data)}
\item{Qvec}{NULL (not available for real data)}
\item{Fvec}{NULL (not available for real data)}
\item{Mvec}{NULL (not available for real data)}
\item{Nvec}{NULL (not available for real data)}
}
}
\usage{
real_data
}
\description{
Real observational data from external source
}
\keyword{datasets}
