% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reduce_Bvec}
\alias{reduce_Bvec}
\title{reduce_Bvec
Function for reducing the breakpoint vector Bvec as result from accepting a backward step}
\usage{
reduce_Bvec(Bvec, BR, S)
}
\arguments{
\item{BR}{(const int) index of the breakpoint to be removed}

\item{S}{(const int) number of current segments}

\item{vec}{(NumericVector&) vector with breakpoint}
}
\value{
void no output is returned because the input (vec) is passed by reference and hence modified
}
\description{
reduce_Bvec
Function for reducing the breakpoint vector Bvec as result from accepting a backward step
}
