% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reduce_vecM}
\alias{reduce_vecM}
\title{reduce_vecM
Function for reducing/updating the state vector Mvec as resulting from accepting a backward step}
\usage{
reduce_vecM(vec, value, BR, S)
}
\arguments{
\item{vec}{(NumericVector&) vector to be augmented}

\item{BR}{(const int) index of the removed breakpoint (range (1,S-1) )}

\item{S}{(const int) number of current segments}

\item{(const}{double) state value of the new segment}
}
\value{
void no output is returned because the input (vec) is passed by reference and hence modified
}
\description{
reduce_vecM
Function for reducing/updating the state vector Mvec as resulting from accepting a backward step
}
