% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reduce_vecZQFV}
\alias{reduce_vecZQFV}
\title{reduce_vecZQFV
Function for reducing/updating the state vectors (Zvec, Vvec, Fvec) as resulting from accepting a backward step}
\usage{
reduce_vecZQFV(vec, value, BR, S)
}
\arguments{
\item{vec}{(IntegerVector&) vector to be augmented}

\item{BR}{(const int) index of the removed breakpoint (range (1,S-1) )}

\item{S}{(const int) number of current segments}

\item{(const}{int) state value of the new segment}
}
\value{
void no output is returned because the input (vec) is passed by reference and hence modified
}
\description{
reduce_vecZQFV
Function for reducing/updating the state vectors (Zvec, Vvec, Fvec) as resulting from accepting a backward step
}
