% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resampling_func}
\alias{resampling_func}
\title{systematic_resampling:  function for performing systematic resampling}
\usage{
resampling_func(
  weight_vec,
  container_B,
  container_V,
  container_Z,
  container_Q,
  container_F,
  Svec,
  V_last_complete,
  B_last,
  n_particle
)
}
\arguments{
\item{weight_vec}{vector with the current normalized weight for each particle}

\item{container_B}{List for with Bvec generated inside the current Update Interval, for each particle}

\item{container_V}{List for with Vvec generated inside the current Update Interval, for each particle}

\item{container_Z}{List for with Zvec generated inside the current Update Interval, for each particle}

\item{container_Q}{List for with Qvec generated inside the current Update Interval, for each particle}

\item{container_F}{List for with Fvec generated inside the current Update Interval, for each particle}

\item{Svec}{vector with number of segment in each particle}

\item{V_last_complete}{Vector with the V state for the last complete segment of each particle}

\item{B_last}{(NumericVector&) Vector with the last simulated changepoint WITHIN the Update Interval, at the previous SMC iteration}

\item{n_particle}{number of particles created}
}
\value{
The function does not explicitly return any object: However, by reference 9 element are updated:
}
\description{
systematic_resampling:  function for performing systematic resampling
}
