% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RJMCMC_SMC}
\alias{RJMCMC_SMC}
\title{RJMCMC_SMC:  function for generating samples from the important density through RJMCMC updates}
\usage{
RJMCMC_SMC(
  T_seg,
  Y_seg,
  U,
  K,
  W,
  start_point,
  end_point,
  t_star,
  num_logs,
  lambdamat,
  keyvec,
  etavec,
  key0vec,
  eta0vec,
  alphavec,
  muvec,
  probvec_V,
  probvec_Z,
  probvec_Q,
  probvec_F,
  P0,
  minimum_n,
  Jss1,
  Js1s,
  Smax,
  n_ite,
  burn_in,
  thinning,
  n_particle,
  particle_index_vec,
  V_last_complete,
  B_last,
  container_B,
  container_V,
  container_Z,
  container_Q,
  container_F,
  Svec,
  weight_vec,
  empty_seg
)
}
\arguments{
\item{T_seg}{Vector with time stamps within the update interval}

\item{Y_seg}{Vector with the messages within the update interval}

\item{U}{Number of levels of the V state}

\item{K}{Number of levels of the Z state}

\item{W}{Number of levels of the Q state}

\item{start_point}{start point of the update interval}

\item{end_point}{(const double&) end point of the update interval}

\item{t_star}{Estimate of the latest changepoint simulated when performing the RJMCMC during the previous iteration of the SMC (precedent update interval)}

\item{num_logs}{Total number of different messages that can be observed (dictionary)}

\item{lambdamat}{matrix U*num_logs with probability mass for the messages in each V state}

\item{keyvec}{Shape parameters ruling the length of non-empty segments in different levels of Z state}

\item{etavec}{Mean parameters ruling the length of non-empty segments in different levels of Z state}

\item{key0vec}{Shape parameters ruling the length of empty segments in different levels of F state}

\item{eta0vec}{Mean parameters ruling the length of empty segments in different levels of F state}

\item{alphavec}{Shape parameters ruling the rate of occurrence in different levels of Z state (non-empty segments)}

\item{muvec}{Mean parameters ruling the rate of occurrence in different levels of Z state}

\item{probvec_V}{probability mass for the V state}

\item{probvec_Z}{probability mass for the Z state}

\item{probvec_Q}{probability mass for the Q state}

\item{probvec_F}{probability mass for the F state}

\item{P0}{(const double&) Probability to observe an empty segment after a non-empty one}

\item{minimum_n}{minimum number of observations that must be observed in a non-empty segment}

\item{Jss1, }{(const double&) Probability to propose a jump forward}

\item{Js1s, }{(const double&) Probability to propose a jump backward}

\item{Smax}{maximum number of segments allowed inside the update interval}

\item{n_ite}{(const int&), number of iteration to be performed in the RJMCMC}

\item{burn_in}{(const int&) Number of iterations used as burn_in period.}

\item{thinning}{(const int&) 1 every "thinning" iterations will be used}

\item{n_particle}{number of particles created}

\item{particle_index_vec}{vector with the index of the particles that will be updated}

\item{V_last_complete}{Vector with the V state for the last complete segment of each particle}

\item{B_last}{(NumericVector&) Vector with the last simulated changepoint WITHIN the Update Interval, at the previous SMC iteration}

\item{container_B}{List for with Bvec generated inside the current Update Interval, for each particle}

\item{container_V}{List for with Vvec generated inside the current Update Interval, for each particle}

\item{container_Z}{List for with Zvec generated inside the current Update Interval, for each particle}

\item{container_Q}{List for with Qvec generated inside the current Update Interval, for each particle}

\item{container_F}{List for with Fvec generated inside the current Update Interval, for each particle}

\item{Svec}{vector with number of segment in each particle}

\item{weight_vec}{vector with the current normalized weight for each particle}

\item{empty_seg}{(Logical) if true then the importance density for empty segments is used}
}
\value{
The function does not explicitly return any object: However, by reference 9 element are updated:
\describe{
\item{container_B}{List for with Bvec generated inside the current Update Interval, for each particle}
\item{container_V}{List for with Vvec generated inside the current Update Interval, for each particle}
\item{container_Z}{List for with Zvec generated inside the current Update Interval, for each particle}
\item{container_Q}{List for with Qvec generated inside the current Update Interval, for each particle}
\item{container_F}{List for with Fvec generated inside the current Update Interval, for each particle}
\item{Svec}{vector with number of segment in each particle}
\item{B_last}{Vector with the last simulated changepoint WITHIN the Update Interval, at the previous SMC iteration}
\item{V_last_complete}{Vector with the V state for the last complete segment of each particle}
\item{weight_vec}{vector with the updated  non normalized weight for each particle}
}
}
\description{
RJMCMC_SMC:  function for generating samples from the important density through RJMCMC updates
}
