% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{segment_box}
\alias{segment_box}
\title{center_segment_box
Function for computing all the necessary quantities needed for updating/evaluating a CLOSED segment or an OPEN segment with at least 1 obs inside.
This function will either sample new states V,Z,Q,F or use the current ones to evaluate the joint proposal and
and the joint posterior distribution for these values}
\usage{
segment_box(
  T_seg,
  Y_seg,
  V,
  Z,
  Q,
  F,
  U,
  K,
  W,
  LB,
  UB,
  num_logs,
  lambdamat,
  keyvec,
  etavec,
  key0vec,
  eta0vec,
  alphavec,
  muvec,
  probvec_V,
  probvec_Z,
  probvec_Q,
  probvec_F,
  V_left,
  P0,
  minimum_n,
  open_segment = FALSE,
  end_point = 0,
  sample_state = FALSE
)
}
\arguments{
\item{T_seg}{Vector with time stamps within the update interval}

\item{Y_seg}{Vector with the messages within the update interval}

\item{V}{(const int&) Value of the current V if present; if not set it 0.}

\item{Z}{(const int&) Value of the current Z if present; if not set it 0.}

\item{Q}{(const int&) Value of the current Q if present; if not set it 0.}

\item{F}{(const int&) Value of the current F if present; if not set it 0.}

\item{U}{(const int&) Number of levels for the V state}

\item{K}{(const int&) Number of levels of the Z state}

\item{W}{(const int&) Number of levels for the Q state}

\item{LB}{(const double&) lower bound of the selected segment}

\item{UB}{(const double&) upper bound of the selected segment}

\item{num_logs}{(const int&) Total number of unique messages that can be observed}

\item{lambdamat}{(const NumericMatrix&) probability weights ruling the message composition of the segments for the different levels of the V state}

\item{keyvec}{(const NumericVector&) Shape parameters ruling the length of non-empty segments in different levels of Z state}

\item{etavec}{(const NumericVector&) Mean parameters ruling the length of non-empty segments in different levels of Z state}

\item{key0vec}{(const NumericVector&) Shape parameters ruling the length of empty segments in different levels of F state}

\item{eta0vec}{(const NumericVector&) Mean parameters ruling the length of empty segments in different levels of F state}

\item{alphavec}{(const NumericVector&) Shape parameters ruling the rate of occurrence of segment for the different levels of the Q state}

\item{muvec}{(const NumericVector&) Mean parameters ruling the rate of occurrence of segment for the different levels of the Q state}

\item{probvec_V}{(const NumericVector&) Probability mass for the V state distribution}

\item{probvec_Z}{(const NumericVector&) probability mass for the Z state}

\item{probvec_Q}{(const NumericVector&) Probability mass for the Q state distribution}

\item{probvec_F}{(const NumericVector&) probability mass for the F state}

\item{V_left}{(const int&) V value for the previous segment}

\item{P0}{(const double&) Probability to observe an empty segment after a non-empty one}

\item{minimum_n}{(const int&) minimum number of observations that must be observed in a non-empty segment}

\item{open_segment}{(const bool, default is false) if false  the segment is considered totally observed. If true the semgent is considered partially observed; this affects the likelihood computation}

\item{end_point}{(double, default is 0.0) to be used for partially observed segments; declares the point where the stops to be observed}

\item{sample_state}{(const bool=false) if true new V,Z,Q,F are sampled and the full conditional are evaluated at the new values. If false (by default) the full condititionals are evaluated at the input provided}
}
\value{
List with 6 elements
\describe{
\item{V}{int,Value of the current or new V}
\item{Z}{int,Value of the current or new Z}
\item{Q}{int,Value of the current or new Q}
\item{F}{int,Value of the current or new F}
\item{sum_proposals}{double,sum of the log proposal for the current or new values of V, Z, Q and F }
\item{out_posterior_evaluation}{double, joint log posterior distribuion for the quantities related to the selected segment }
}
}
\description{
center_segment_box
Function for computing all the necessary quantities needed for updating/evaluating a CLOSED segment or an OPEN segment with at least 1 obs inside.
This function will either sample new states V,Z,Q,F or use the current ones to evaluate the joint proposal and
and the joint posterior distribution for these values
}
