% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shift_function}
\alias{shift_function}
\title{shift_function: This function takes in the current state of the MCMC and employ a position shift of a breakpoint.
all the input parameters described below refers to their most recent values.}
\usage{
shift_function(
  S,
  Tvec,
  Yvec,
  U,
  K,
  W,
  probvec_V,
  probvec_Z,
  probvec_Q,
  probvec_F,
  lambdamat,
  keyvec,
  etavec,
  key0vec,
  eta0vec,
  alphavec,
  muvec,
  num_logs,
  Bvec,
  Zvec,
  Qvec,
  Vvec,
  Fvec,
  Jss1,
  Js1s,
  P0,
  minimum_n,
  start_point = 0,
  end_point = 0
)
}
\arguments{
\item{S}{(int). The total number of interarrivals (intervals).}

\item{Tvec}{(const NumericVector). Vector of the observed time stamps in the update interval}

\item{Yvec}{(const IntegerVector). Vector of the observed messages in the update interval}

\item{U}{(const int&) Number of levels for the V state}

\item{K}{(const int&) Number of levels of the Z state}

\item{W}{(const int&) Number of levels for the Q state}

\item{probvec_V}{(const NumericVector&) Probability mass for the V state distribution}

\item{probvec_Z}{(const NumericVector&) probability mass for the Z state}

\item{probvec_Q}{(const NumericVector&) Probability mass for the Q state distribution}

\item{probvec_F}{(const NumericVector&) probability mass for the F state}

\item{lambdamat}{(const NumericMatrix&) probability weights ruling the message composition of the segments for the different levels of the V state}

\item{keyvec}{(const NumericVector&) Shape parameters ruling the length of non-empty segments in different levels of Z state}

\item{etavec}{(const NumericVector&) Mean parameters ruling the length of non-empty segments in different levels of Z state}

\item{key0vec}{(const NumericVector&) Shape parameters ruling the length of empty segments in different levels of F state}

\item{eta0vec}{(const NumericVector&) Mean parameters ruling the length of empty segments in different levels of F state}

\item{alphavec}{(const NumericVector&) Shape parameters ruling the rate of occurrence of segment for the different levels of the Q state}

\item{muvec}{(const NumericVector&) Mean parameters ruling the rate of occurrence of segment for the different levels of the Q state.}

\item{num_logs}{(const int). Total numer of different logs (levels of the factor )}

\item{Bvec}{(NumericVector). Current vector of changepoints}

\item{Zvec}{(IntegerVector). Current vector with Z state of each  segment}

\item{Qvec}{(IntegerVector). Current vector with Q state of each  segment}

\item{Vvec}{(IntegerVector). Current vector with V state of each  segment}

\item{Fvec}{(IntegerVector). Current vector with F state of each  segment}

\item{Jss1}{(const double). Probability to propose a forward move}

\item{Js1s}{(const double). Probability to propose a backward move}

\item{P0}{(const double&) probability of observing a non-empty segment after a empty one}

\item{minimum_n}{(const int&) minimum number of observations in a non-empty segment}

\item{start_point}{(double, default is 0.0) start of the update interval}

\item{end_point}{(double, default is 0.0) end of the update interval; to be used for partially observed segments}
}
\description{
shift_function: This function takes in the current state of the MCMC and employ a position shift of a breakpoint.
all the input parameters described below refers to their most recent values.
}
