% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SMC_turcotte_cpp}
\alias{SMC_turcotte_cpp}
\title{SMC Function implementing particle filter for the model in  Gramuglia et al. 2020 using the importance density approximation
as in Heard & Turcotte (2017)}
\usage{
SMC_turcotte_cpp(
  Yvec,
  Tvec,
  length_UI,
  n_particle,
  U,
  W,
  K,
  num_logs,
  lambdamat,
  keyvec,
  etavec,
  key0vec,
  eta0vec,
  alphavec,
  muvec,
  probvec_V,
  probvec_Z,
  probvec_Q,
  probvec_F,
  P0,
  minimum_n,
  Jss1,
  Js1s,
  Smax,
  n_ite,
  burn_in,
  thinning
)
}
\arguments{
\item{Yvec}{IntegerVector  with all the messages}

\item{Tvec}{NumericVector with all the time-stamps}

\item{length_UI}{length of the each update interval}

\item{n_particle}{number of particles created}

\item{U}{Number of levels of the V state}

\item{W}{Number of levels of the Q state}

\item{K}{Number of levels of the Z state}

\item{num_logs}{Total number of different messages that can be observed (dictionary)}

\item{lambdamat}{matrix U*num_logs with probability mass for the messages in each V state}

\item{keyvec}{Shape parameters ruling the length of non-empty segments in different levels of Z state}

\item{etavec}{Mean parameters ruling the length of non-empty segments in different levels of Z state}

\item{key0vec}{Shape parameters ruling the length of empty segments in different levels of F state}

\item{eta0vec}{Mean parameters ruling the length of empty segments in different levels of F state}

\item{alphavec}{Shape parameters ruling the rate of occurrence in different levels of Z state (non-empty segments)}

\item{muvec}{Mean parameters ruling the rate of occurrence in different levels of Z state}

\item{probvec_V}{probability mass for the V state}

\item{probvec_Z}{probability mass for the Z state}

\item{probvec_Q}{probability mass for the Q state}

\item{probvec_F}{probability mass for the F state}

\item{P0}{(const double&) Probability to observe an empty segment after a non-empty one}

\item{minimum_n}{minimum number of observations that must be observed in a non-empty segment}

\item{Jss1, }{(const double&) Probability to propose a jump forward}

\item{Js1s, }{(const double&) Probability to propose a jump backward}

\item{Smax}{maximum number of segments allowed inside the update interval}

\item{n_ite}{(const int&), number of iteration to be performed in the RJMCMC}

\item{burn_in}{(const int&) Number of iterations used as burn_in period.}

\item{thinning}{(const int&) 1 every "thinning" iterations will be used}
}
\description{
Given a time series of time-stamp and messages this function run a particle filter with
Reversible Jump Markov Chain Montecarlo steps for approximating the posterior distribution for the following variables
\itemize{
\item the number of segments
\item the position of the breakpoints defining the segments
\item the state dynamic for each interval which include
\itemize{
\item Z, ruling the length of each non-empty segment
\item Q, ruling the rate of occurrence of the messages in non-empty segments
\item V, ruling the segment composition (in term of messages) in non-empty segments
\item F, ruling the length of empty-segments
}
}
The algorithm approximate these posterior distributions by splitting the observation
time interval in sub-intervals defined by the custom parameter \code{length_UI}
}
