% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RJSMC.R
\name{plot.RJSMC}
\alias{plot.RJSMC}
\title{plot.RJSMC:  Plotting results from RJSMC::SMC routine}
\usage{
\method{plot}{RJSMC}(
  obj,
  origin = as.Date("2022-01-01"),
  pl = "V",
  truth = list(B = Bvec, cl = Vvec)
)
}
\arguments{
\item{obj}{Object of class RJSMC, the output from RJSMC::SMC}

\item{origin}{Starting date for data}

\item{pl}{Equal to V, Z, Q or F depending on which variable to pl}
}
\description{
Plots an area plot over time with the areas corresponding to probabilities for the different discrete events
}
