% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJSMC_class.R
\docType{class}
\name{RJSMC-class}
\alias{RJSMC-class}
\alias{RJSMC}
\title{An S4 class to represent the output of the SMC function.}
\description{
An S4 class to represent the output of the SMC function.
}
\section{Slots}{

\describe{
\item{\code{n_UI}}{number of Update Intervals within which the SMC has been performed}

\item{\code{points_container}}{list containing \code{n_UI} vector: each vector represent the time point where the marginal posterior distributions for the state variables have been approximated, for each different Update Interval}

\item{\code{posteriors_container_V}}{list containing \code{n_UI} matrices: each matrix represent the marginal posterior distribution of the state variable \code{V}, evaluated at each discretization point (stored in \code{point_containers}), for each different Update interval}

\item{\code{posteriors_container_Z}}{list containing \code{n_UI} matrices: each matrix represent the marginal posterior distribution of the state variable \code{Z}, evaluated at each discretization point (stored in \code{point_containers}), for each different Update interval}

\item{\code{posteriors_container_Q}}{list containing \code{n_UI} matrices: each matrix represent the marginal posterior distribution of the state variable \code{Q}, evaluated at each discretization point (stored in \code{point_containers}), for each different Update interval}

\item{\code{posteriors_container_F}}{\itemize{
\item list containing \code{n_UI} matrices: each matrix represent the marginal posterior distribution of the state variable \code{F}, evaluated at each discretization point (stored in \code{point_containers}), for each different Update interval A length-one numeric vector
}}

\item{\code{posteriors_container_B}}{numeric vector containing the marginal posterior probability that a breakpoint occurs at each discretization point (stored in \code{points_container}), concatenated across all Update Intervals}

\item{\code{storage_B}}{list containing breakpoints for each particle in each Update Interval (for evaluation purposes)}

\item{\code{storage_V}}{list containing state variable \code{V} for each particle in each Update Interval (for evaluation purposes)}

\item{\code{storage_weight}}{list containing particle weights for each Update Interval (for evaluation purposes)}

\item{\code{UI_index_vector}}{Index vector describing which interval each time points is evaluated within.}

\item{\code{UI_bounds}}{Numeric vector containing the bounds of the Update Intervals.}

\item{\code{elapsed_time}}{Numeric scalar: time in seconds taken by the SMC C++ algorithm only (excludes post-processing).}
}}

