% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{SMC}
\alias{SMC}
\title{SMC Function implementing particle filter for the model in  Gramuglia et al. 2020}
\usage{
SMC(ts_data, parameters, settings)
}
\arguments{
\item{ts_data}{list with data:
\itemize{
\item {Yvec}{IntegerVector  with the observed messages}
\item {Tvec}{NumericVector with  the  observed time-stamps}
}}

\item{parameters}{list with model parameters:
\itemize{
\item {U}{Number of levels of variable V}
\item {W}{Number of levels of variable Q}
\item {K}{Number of levels of variable Z}
\item {lambdamat}{matrix (U * num_logs) hosting the  probability mass for the messages in each V state}
\item {keyvec}{Shape parameters ruling the length of non-empty segments in different levels of Z state}
\item {etavec}{Mean parameters ruling the length of non-empty segments in different levels of Z state}
\item {key0vec}{Shape parameters ruling the length of empty segments in different levels of F state}
\item {eta0vec}{Mean parameters ruling the length of empty segments in different levels of F state}
\item {alphavec}{Shape parameters ruling the rate of occurrence in different levels of Z state (non-empty segments)}
\item {muvec}{Mean parameters ruling the rate of occurrence in different levels of Z state}
\item {probvec_V}{probability mass for the V state}
\item {probvec_Z}{probability mass for the Z state}
\item {probvec_Q}{probability mass for the Q state}
\item {probvec_F}{probability mass for the F state}
\item {P0}{(const double&) Probability to observe an empty segment after a non-empty one}
\item {minimum_n}{minimum number of observations that must be observed in a non-empty segment}
}}

\item{settings}{list with simulation settings parameters:
\itemize{
\item {length_UI}{length of the each update interval}
\item {n_particle}{number of particles to generate}
\item {num_logs}{Total number of different messages that can be observed}
\item {Jss1}{(const double&) Probability to propose a jump forward}
\item {Js1s}{(const double&) Probability to propose a jump backward}
\item {Smax}{maximum number of segments allowed inside the update interval}
\item {n_ite}{(const int&), number of iteration to be performed in the RJMCMC}
\item {burn_in}{(const int&) Number of iterations used as burn_in period.}
\item {thinning}{(const int&) One every "thinning" iterations will be stored by the RJMCMC}
\item {method}{(string}) set it to "turcotte" for applying the turcotte method for importance density approximation)
}}
}
\value{
The output of the function is a list with the following elements
\itemize{
\item n_UI - number of Update Intervals within which the SMC has been performed
\item points_container - vector containing all the time point (i.e. discretization points) where the marginal posterior distributions for the state variables have been approximated
\item posteriors_container_V - matrix (length(points_container) X U) containing the marginal probability mass for state variable V, evaluated at each discretization point
\item posteriors_container_Z - matrix (length(points_container) X K) containing the marginal probability mass for state variable Z, evaluated at each discretization point
\item posteriors_container_Q - matrix (length(points_container) X W) containing the marginal probability mass for state variable Q, evaluated at each discretization point
\item posteriors_container_F - matrix (length(points_container) X 2) containing the marginal probability mass for state variable F, evaluated at each discretization point
\item UI_index_vector - vector with the Update Interval number each discretization point in "points_container" belongs to
}
}
\description{
Given an time series of time-stamps and messages this function run a particle filter with
Reversible Jump Markov Chain Montecarlo steps for approximating the posterior distribution of the following state dynamics variables
\itemize{
\item Z, ruling the length of each non-empty segment
\item Q, ruling the rate of occurrence of the messages in non-empty segments
\item V, ruling the segment composition (in term of messages) in non-empty segments
\item F, ruling the length of empty-segments
}
The algorithm approximate these posterior distributions by splitting the observation
time interval in sub-intervals defined by the custom parameter \code{length_UI}
}
