% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weigths_correction}
\alias{weigths_correction}
\title{weights_correction
Function for computing the weights correction based on the importance density for a segment}
\usage{
weigths_correction(
  T_seg,
  break_sample_output,
  M,
  K,
  keyvec,
  etavec,
  key0vec,
  eta0vec,
  probvec_Z,
  probvec_F,
  P0
)
}
\arguments{
\item{T_seg}{Vector with time stamps within the update interval}

\item{break_sample_output}{(const List&) List of breakpoint vectors generated with the Importance Sampling}

\item{M}{(const int&) number of iteration performed in the Importance Sampling}

\item{K}{(const int&) Number of levels of the Z state}

\item{keyvec}{(const NumericVector&) Shape parameters ruling the length of non-empty segments in different levels of Z state}

\item{etavec}{(const NumericVector&) Mean parameters ruling the length of non-empty segments in different levels of Z state}

\item{key0vec}{(const NumericVector&) Shape parameters ruling the length of empty segments in different levels of F state}

\item{eta0vec}{(const NumericVector&) Mean parameters ruling the length of empty segments in different levels of F state}

\item{probvec_Z}{(const NumericVector&) probability mass for the Z state}

\item{probvec_F}{(const NumericVector&) probability mass for the F state}

\item{P0}{(const double&) Probability to observe an empty segment after a non-empty one}
}
\value{
correction weight (double)
}
\description{
weights_correction
Function for computing the weights correction based on the importance density for a segment
}
